+++ src/runtime/components/InputNumber.vue
@@ -231,7 +231,7 @@ defineExpose({
           <B24Button
             :icon="incrementIcon"
             color="air-tertiary-no-accent"
-            :size="size"
+            :size="inputSize"
             :aria-label="t('inputNumber.increment')"
             v-bind="typeof increment === 'object' ? increment : undefined"
           />
@@ -245,7 +245,7 @@ defineExpose({
           <B24Button
             :icon="decrementIcon"
             color="air-tertiary-no-accent"
-            :size="size"
+            :size="inputSize"
             :aria-label="t('inputNumber.decrement')"
             v-bind="typeof decrement === 'object' ? decrement : undefined"
           />